Separated reducers and actions
